Add length and width fields
Rather than just surface area, which requires that the code wrangling
the desk compute and pass a new surface area, the methods to set length
and width both automatically call to change surface area when they are
invoked. Both of these methods invoke the change to surface area, which
allows for editing either of them individually.

Since the structure of their individual methods is exactly the same, I'm
tempted to run with a different solution. Ideally, one would call to
change a dimension, passing the name of the dimension and the new value,
but the means of doing this dynamically in java are beyond my knowledge
at this time. Probably involves the java equivalent of exec or somesuch,
which is almost always a dangerous can of worms to open. This works for
now, and I can always change it later if I need to.