Change set_author to void return type
The convention I've witnessed uses a void return on mutators. I
certainly think it would be useful to return an object having been
modified, such as a number being incremented in some way, but I suppose
when the change is simple and the only thing being returned would be
exactly what got passed in, it doesn't add much value.