Add mutator and accessor for isbn
ISBN is nominally a number, but its format also includes dashes. This
could be implemented with some long primitive type nonsense, but that
involved excessive cleverness for the purposes here. Even for this
purpose, I'm not verifying the length or building a regex for it. This
is both an alpha and an assignment, meant to demonstrate understanding
of structure and logic, rather than the exact formatting requirements of
a real-world standard.